requirements.txt
# lib
numpy==1.22.1
scipy
torch==1.11.0
torchvision==0.12.0
timm==0.6.7
faiss-gpu
pillow
lmdb
pandas==1.4.1
scikit-learn==1.0.2
tqdm==4.62.3
tensorboard
pytorch-metric-learning
hydra-core>=1.1.1
hydra_colorlog>=1.1.0
hydra-ray-launcher
hydra_torchrun_launcher @ git+https://github.com/facebookresearch/hydra.git#egg=hydra_torchrun_launcher&subdirectory=contrib/hydra_torchrun_launcher

# utils
matplotlib==3.5.1
seaborn==0.11.2
SciencePlots
plotext
jupyter
ipython
gdown

# dev
ipdb
flake8
flake8-annotations
autopep8
pylint
isort
pytest
